Arreglado el buscador de Eventos y cambio de localización de la llamada de la api de eventos
