+++ src/pages/Eventos.jsx
@@ -1,45 +1,9 @@
 import GrupoDeCardEvento from "../components/cards/GrupoDeCardEvento";
 import SearchBar from "../components/common/SearchBar";
 import Calendario from "../components/common/Calendario";
-import { useState, useEffect } from "react";
-import clienteAxios from "../config/axios"
-import { formatearFechaHora } from "../utilities/formatearFechaHora"
+import { useState } from "react";
 
-export default function Eventos() {
-
-    const [datos, setDatos] = useState([])
-
-    useEffect(() => {
-        const obtenerDatos = async () => {
-            try {
-                const response = await clienteAxios.get('/eventos/activos');
-                const eventosFormateados = response.data.map(item => {
-                    const { fecha, hora } = formatearFechaHora(item[4]);
-
-                    return {
-                        id: item[0],
-                        titulo: item[1],
-                        imagen: item[2],
-                        descripcionEvento: item[3],
-                        fechaDisplay: fecha,
-                        horaDisplay: hora,
-                        categoria: item[5],
-                        descripcionCategoria: item[6],
-                        material: item[7],
-                        ubicacion: item[8],
-                        participantes: item[9]
-                    };
-                });
-
-                setDatos(eventosFormateados);
-            } catch (error) {
-                console.error('Error al obtener los datos:', error);
-            }
-        };
-
-        obtenerDatos();
-    }, []);
-
+export default function Eventos({ datos }) {
 
     const [query, setQuery] = useState("");
 
@@ -50,7 +14,7 @@ export default function Eventos() {
 
     const filteredEvents = datos.filter((e) => {
         const text = query.toLowerCase();
-        const fechaEvento = new Date(e.fecha);
+        const fechaEvento = new Date(e.fechaDisplay);
 
         const cumpleTexto = e.titulo.toLowerCase().includes(text);
